perf: Change size_t into uint_32t

+++ chain.h
@@ -33,7 +33,7 @@ namespace original {
                 static void connect(chainNode* prev, chainNode* next);
         };
 
-        size_t size_;
+        uint32_t size_;
         chainNode* begin_;
         chainNode* end_;
 
@@ -58,11 +58,11 @@ namespace original {
         chain& operator=(const chain& other);
         bool operator==(const chain& other) const;
         void operator+=(chain& other);
-        [[nodiscard]] size_t size() const override;
+        [[nodiscard]] uint32_t size() const override;
         TYPE get(int64_t index) const override;
         TYPE& operator[](int64_t index) override;
         void set(int64_t index, const TYPE &e) override;
-        size_t indexOf(const TYPE &e) const override;
+        uint32_t indexOf(const TYPE &e) const override;
         void pushBegin(const TYPE &e) override;
         void push(int64_t index, const TYPE &e) override;
         void pushEnd(const TYPE &e) override;
@@ -208,7 +208,7 @@ namespace original {
     template <typename TYPE>
     original::chain<TYPE>::chain(array<TYPE> arr)
         : chain() {
-        for (size_t i = 0; i < arr.size(); i++) {
+        for (uint32_t i = 0; i < arr.size(); i++) {
             auto* cur_node = new chainNode(arr.get(i));
             if (this->size() == 0)
             {
@@ -256,7 +256,7 @@ namespace original {
         if (this->size() != other.size()) return false;
         auto* this_ = this->begin_;
         auto* other_ = other.begin_;
-        for (size_t i = 0; i < this->size(); ++i) {
+        for (uint32_t i = 0; i < this->size(); ++i) {
             if (this_->getVal() != other_->getVal()){
                 return false;
             }
@@ -278,7 +278,7 @@ namespace original {
     }
 
     template <typename TYPE>
-    auto original::chain<TYPE>::size() const -> size_t
+    auto original::chain<TYPE>::size() const -> uint32_t
     {
         return this->size_;
     }
@@ -299,13 +299,13 @@ namespace original {
         chainNode* cur;
         if (!reverse_visit){
             cur = this->begin_;
-            for(size_t i = 0; i < this->parseNegIndex(index); i++)
+            for(uint32_t i = 0; i < this->parseNegIndex(index); i++)
             {
                 cur = cur->getPNext();
             }
         } else{
             cur = this->end_;
-            for(size_t i = this->size() - 1; i > this->parseNegIndex(index); i-= 1)
+            for(uint32_t i = this->size() - 1; i > this->parseNegIndex(index); i-= 1)
             {
                 cur = cur->getPPrev();
             }
@@ -323,13 +323,13 @@ namespace original {
         chainNode* cur;
         if (!reverse_visit){
             cur = this->begin_;
-            for(size_t i = 0; i < this->parseNegIndex(index); i++)
+            for(uint32_t i = 0; i < this->parseNegIndex(index); i++)
             {
                 cur = cur->getPNext();
             }
         } else{
             cur = this->end_;
-            for(size_t i = this->size() - 1; i > this->parseNegIndex(index); i-= 1)
+            for(uint32_t i = this->size() - 1; i > this->parseNegIndex(index); i-= 1)
             {
                 cur = cur->getPPrev();
             }
@@ -348,13 +348,13 @@ namespace original {
         chainNode* cur;
         if (!reverse_visit){
             cur = this->begin_;
-            for(size_t i = 0; i < this->parseNegIndex(index); i++)
+            for(uint32_t i = 0; i < this->parseNegIndex(index); i++)
             {
                 cur = cur->getPNext();
             }
         } else{
             cur = this->end_;
-            for(size_t i = this->size() - 1; i > this->parseNegIndex(index); i -= 1)
+            for(uint32_t i = this->size() - 1; i > this->parseNegIndex(index); i -= 1)
             {
                 cur = cur->getPPrev();
             }
@@ -367,8 +367,8 @@ namespace original {
     }
 
     template <typename TYPE>
-    auto original::chain<TYPE>::indexOf(const TYPE &e) const -> size_t {
-        size_t i = 0;
+    auto original::chain<TYPE>::indexOf(const TYPE &e) const -> uint32_t {
+        uint32_t i = 0;
         for (chainNode* current = this->begin_; current != nullptr; current = current->getPNext()) {
             if (current->getVal() == e) {
                 return i;
@@ -378,7 +378,7 @@ namespace original {
         return this->size();
     }
 
-template <typename TYPE>
+    template <typename TYPE>
     auto original::chain<TYPE>::pushBegin(const TYPE &e) -> void
     {
         auto* new_node = new chainNode(e);
@@ -409,13 +409,13 @@ template <typename TYPE>
             chainNode* cur;
             if (!reverse_visit){
                 cur = this->begin_;
-                for(size_t i = 0; i < index; i++)
+                for(uint32_t i = 0; i < index; i++)
                 {
                     cur = cur->getPNext();
                 }
             } else{
                 cur = this->end_;
-                for(size_t i = this->size() - 1; i > index; i -= 1)
+                for(uint32_t i = this->size() - 1; i > index; i -= 1)
                 {
                     cur = cur->getPPrev();
                 }
