+++ container.h
@@ -8,7 +8,7 @@ namespace original {
     template <typename TYPE>
     class container{
     public:
-        [[nodiscard]] virtual size_t size() const = 0;
+        [[nodiscard]] virtual uint32_t size() const = 0;
         [[nodiscard]] bool empty() const;
         virtual bool contains(const TYPE& e) const = 0 ;
         virtual void add(const TYPE& e) = 0;
